requirements.txt
contourpy==1.3.1 ; python_version >= "3.12"
cycler==0.12.1 ; python_version >= "3.12"
fonttools==4.55.7 ; python_version >= "3.12"
kiwisolver==1.4.8 ; python_version >= "3.12"
matplotlib==3.10.0 ; python_version >= "3.12"
numpy==2.2.2 ; python_version >= "3.12"
packaging==24.2 ; python_version >= "3.12"
pandas==2.2.3 ; python_version >= "3.12"
pillow==11.1.0 ; python_version >= "3.12"
pyparsing==3.2.1 ; python_version >= "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.12"
pytz==2024.2 ; python_version >= "3.12"
six==1.17.0 ; python_version >= "3.12"
tk==0.1.0 ; python_version >= "3.12"
tzdata==2025.1 ; python_version >= "3.12"
scipy==1.15.1 ; python_version >= "3.12"
opencv-python-headless==4.11.0.86 ; python_version >= "3.12"